Corrected System of Equations
Reaction 1 is not linearly independent so it was removed by setting e1 =0.
Didn't wanna manually delete anything so e2 and e3 are still numbered as such but they are root 1 and root 2 of the system respectively.
Two new inputs were created to account for the desired feed ratio of CO2 to H2.